Split field models into separate files
